refactor(User,App, Auth): adding middleware and actions to ensure the credential data is fetched on user auth
