+++ src/api/users/users.constants.ts
@@ -1,7 +1,6 @@
-import { mergeLeft, mergeRight, pipe } from 'ramda'
+import { mergeRight } from 'ramda'
 
 import { ApiClients, ApiMeta, ApiMethods } from '../api.types'
-import { addIdBeforeEndpointInConfig } from '../api.utils'
 import { UsersEndpoints } from './users.types'
 
 export const USERS_CONFIG: Partial<ApiMeta> = {
@@ -15,12 +14,10 @@ export const USERS_EDIT_CONFIG: Partial<ApiMeta> = mergeRight(USERS_CONFIG, {
 })
 export const USERS_GET_BY_ID_CONFIG: Partial<ApiMeta> = mergeRight(USERS_CONFIG, {})
 
-export const USERS_CREDENTIALS_GET_BY_ID_CONFIG: Partial<ApiMeta> = pipe(
-  mergeLeft({
-    endpoint: UsersEndpoints.Credentials,
-  }),
-  addIdBeforeEndpointInConfig,
-)(USERS_CONFIG)
+export const USERS_CREDENTIALS_GET_BY_ID_CONFIG: Partial<ApiMeta> = mergeRight(USERS_CONFIG, {
+  endpoint: UsersEndpoints.Credentials,
+})
+
 export const USERS_CREDENTIALS_GET_BY_TYPE_CONFIG: Partial<ApiMeta> = mergeRight(USERS_CONFIG, {
   endpoint: UsersEndpoints.Credentials,
 })
