+++ src/modules/App/App.middleware.ts
@@ -1,7 +1,8 @@
 import { Middleware } from 'redux'
 
-import * as AuthActions from './../Auth/Auth.reducer'
-import { resetAppData } from './App.reducer'
+import { actions as AuthActions } from '../Auth'
+import { actions as UserActions } from '../User'
+import { fetchInitial, resetAppData } from './App.reducer'
 
 export const appResetFlow: Middleware =
   ({ dispatch }) =>
@@ -13,5 +14,16 @@ export const appResetFlow: Middleware =
       dispatch(AuthActions.clearAuth())
     }
 
+    return result
+  }
+
+export const fetchInitialFlow: Middleware =
+  ({ dispatch }) =>
+  next =>
+  action => {
+    const result = next(action)
+    if (fetchInitial.match(action)) {
+      dispatch(UserActions.fetchUserCredentials)
+    }
     return result
   }
