+++ src/modules/App/App.reducer.ts
@@ -2,4 +2,5 @@ import { createAction } from '@reduxjs/toolkit'
 
 const name = '[App]'
 
-export const resetAppData = createAction(`${name} Clear App Data`)
+export const resetAppData = createAction(`${name} resetAppData`)
+export const fetchInitial = createAction(`${name} fetchInitial`)
